término de funcionalidad para implementar noticias
