+++ app/src/main/java/com/example/currentnews/views/news/DetNewsFragment.kt
@@ -1,24 +1,23 @@
 package com.example.currentnews.views.news
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.example.currentnews.R
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class DetNewsFragment : Fragment() {
 
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
     }
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
