+++ app/src/main/res/layout/fragment_item_news.xml
@@ -10,9 +10,9 @@
 
         <ImageView
             android:id="@+id/img_news"
-            android:layout_width="@dimen/marginTop_login"
-            android:layout_height="@dimen/marginTop_login"
-            android:layout_marginStart="@dimen/commo_padding_min"
+            android:layout_width="@dimen/marginTop_image"
+            android:layout_height="@dimen/marginTop_image"
+            android:layout_margin="@dimen/commo_padding_min"
             android:contentDescription="@string/description_img"
             android:src="@drawable/playa"
             app:layout_constraintStart_toStartOf="parent"
@@ -24,11 +24,11 @@
             style="@style/TextAppearance.AppCompat.Body1"
             android:layout_width="@dimen/marginTop_login"
             android:layout_height="0dp"
-            android:layout_marginTop="10dp"
+            android:layout_marginTop="@dimen/commo_padding_min"
+            android:layout_marginStart="@dimen/commo_padding_middle"
             android:lines="@integer/recipebook_line_min"
             android:text="@string/name_news"
             android:textStyle="bold"
-            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toEndOf="@id/img_news" />
 
@@ -36,15 +36,14 @@
             android:id="@+id/tv_description_news"
             android:layout_width="wrap_content"
             android:layout_height="0dp"
-            android:layout_marginStart="68dp"
-            android:layout_marginTop="36dp"
+            android:layout_marginStart="@dimen/commo_padding_middle"
+            android:layout_marginTop="@dimen/commo_padding_min"
             android:gravity="left"
             android:maxLines="1"
             android:text="@string/author"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toEndOf="@+id/img_news"
-            app:layout_constraintTop_toTopOf="@id/tv_title_News"
-            app:layout_constraintVertical_bias="0.0" />
+            app:layout_constraintTop_toBottomOf="@+id/tv_title_News"/>
 
         <View
             android:layout_width="match_parent"
