+++ app/src/main/res/values/colors.xml
@@ -8,5 +8,7 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="purple">#9C27B0</color>
+    <color name="rose">#E0B7CE</color>
+
     <color name="text_color_reg">#3F51B5</color>
 </resources>
