+++ app/src/main/res/values/dimens.xml
@@ -6,6 +6,7 @@
     <dimen name="commo_padding_middle">24dp</dimen>
 
     <dimen name="marginTop_login">80dp</dimen>
+    <dimen name="marginTop_image">60dp</dimen>
     <dimen name="height_login">40dp</dimen>
     <dimen name="text_Size">20dp</dimen>
     <dimen name="edit_text_width">210dp</dimen>
